+++ source/app/service-providers/fsal/index.ts
@@ -34,7 +34,7 @@ import type { SearchTerm } from '@dts/common/search'
 import ProviderContract from '@providers/provider-contract'
 import { app, ipcMain } from 'electron'
 import type LogProvider from '@providers/log'
-import { hasMarkdownExt, hasCodeExt, MD_EXT } from '@common/util/file-extention-checks'
+import { hasMarkdownExt, hasCodeExt } from '@common/util/file-extention-checks'
 import getMarkdownFileParser from './util/file-parser'
 import type ConfigProvider from '@providers/config'
 import { promises as fs, constants as FS_CONSTANTS } from 'fs'
@@ -197,9 +197,10 @@ export default class FSAL extends ProviderContract {
    * emitted.
    */
   private async syncRoots (): Promise<void> {
-    const { openPaths } = this._config.get()
+    const { openFiles, openWorkspaces } = this._config.get().app
+    const allRoots = openFiles.concat(openWorkspaces)
 
-    for (const rootPath of openPaths) {
+    for (const rootPath of allRoots) {
       if (this.watchers.has(rootPath)) {
         continue // This path has already been loaded
       }
@@ -227,7 +228,7 @@ export default class FSAL extends ProviderContract {
     // Before finishing up, unwatch all roots that are no longer part of the
     // config
     for (const [ rootPath, watcher ] of this.watchers) {
-      if (!openPaths.includes(rootPath)) {
+      if (!allRoots.includes(rootPath)) {
         await watcher.shutdown()
         this.watchers.delete(rootPath)
       }
@@ -244,13 +245,17 @@ export default class FSAL extends ProviderContract {
     // Start a timer to measure how long the roots take to load.
     let start = performance.now()
 
-    const { openPaths } = this._config.get()
+    const { openFiles, openWorkspaces } = this._config.get().app
     const pathsToIndex: string[] = []
-    for (const rootPath of openPaths) {
-      if (await this.isFile(rootPath) && !path.basename(rootPath).startsWith('.')) {
-        pathsToIndex.push(rootPath)
-      } else if (await this.isDir(rootPath) && !ignoreDir(rootPath)) {
-        const allPaths = await this.readDirectoryRecursively(rootPath)
+    for (const file of openFiles) {
+      if (await this.isFile(file) && !path.basename(file).startsWith('.')) {
+        pathsToIndex.push(file)
+      }
+    }
+
+    for (const workspace of openWorkspaces) {
+      if (await this.isDir(workspace) && !ignoreDir(workspace)) {
+        const allPaths = await this.readDirectoryRecursively(workspace)
         pathsToIndex.push(...allPaths)
       }
     }
@@ -294,14 +299,18 @@ export default class FSAL extends ProviderContract {
    * @return  {Promise<AnyDescriptor>[]}  The descriptors
    */
   public async getAllLoadedDescriptors (): Promise<AnyDescriptor[]> {
-    const { openPaths } = this._config.get()
+    const { openFiles, openWorkspaces } = this._config.get().app
     const allDescriptors: AnyDescriptor[] = []
 
-    for (const rootPath of openPaths) {
-      if (await this.isFile(rootPath) && !path.basename(rootPath).startsWith('.')) {
-        allDescriptors.push(await this.getDescriptorFor(rootPath))
-      } else if (await this.isDir(rootPath) && !ignoreDir(rootPath)) {
-        const allPaths = await this.readDirectoryRecursively(rootPath)
+    for (const file of openFiles) {
+      if (await this.isFile(file) && !path.basename(file).startsWith('.')) {
+        allDescriptors.push(await this.getDescriptorFor(file))
+      }
+    }
+
+    for (const workspace of openWorkspaces) {
+      if (await this.isDir(workspace) && !ignoreDir(workspace)) {
+        const allPaths = await this.readDirectoryRecursively(workspace)
         for (const child of allPaths) {
           const descriptor = await this.getDescriptorFor(child)
           allDescriptors.push(descriptor)
@@ -335,10 +344,8 @@ export default class FSAL extends ProviderContract {
         return descriptor
       }
 
-      for (const type of MD_EXT) {
-        if (descriptor.name === query + type) {
-          return descriptor
-        }
+      if (descriptor.name === query + descriptor.ext) {
+        return descriptor
       }
     }
   }
@@ -459,12 +466,12 @@ export default class FSAL extends ProviderContract {
    * @return  {Promise<boolean>}           Returns true, if absPath is a dir
    */
   public async isDir (absPath: string): Promise<boolean> {
-    if (!await this.pathExists(absPath)) {
+    try {
+      const stat = await fs.lstat(absPath)
+      return stat.isDirectory()
+    } catch (err: any) {
       return false
     }
-
-    const metadata = await getFilesystemMetadata(absPath)
-    return metadata.isDirectory
   }
 
   /**
@@ -475,12 +482,12 @@ export default class FSAL extends ProviderContract {
    * @return  {Promise<boolean>}           Returns true, if absPath is a file
    */
   public async isFile (absPath: string): Promise<boolean> {
-    if (!await this.pathExists(absPath)) {
+    try {
+      const stat = await fs.lstat(absPath)
+      return stat.isFile()
+    } catch (err: any) {
       return false
     }
-
-    const metadata = await getFilesystemMetadata(absPath)
-    return metadata.isFile
   }
 
   /**
@@ -670,14 +677,6 @@ export default class FSAL extends ProviderContract {
    * @return  {Promise<string>}           Resolves with a string
    */
   public async loadAnySupportedFile (absPath: string): Promise<string> {
-    if (await this.isDir(absPath)) {
-      throw new Error(`[FSAL] Cannot load file ${absPath} as it is a directory`)
-    }
-
-    if (!await this.isFile(absPath)) {
-      throw new Error(`[FSAL] Cannot load file ${absPath}: Not found`)
-    }
-
     const descriptor = await this.getDescriptorForAnySupportedFile(absPath)
 
     if (descriptor.type === 'file') {
@@ -703,25 +702,22 @@ export default class FSAL extends ProviderContract {
    * @throws if the path is not a file
    */
   public async getDescriptorForAnySupportedFile (absPath: string): Promise<MDFileDescriptor|CodeFileDescriptor|OtherFileDescriptor> {
+    if (await this.isFile(absPath)) {
+      if (hasMarkdownExt(absPath)) {
+        return await FSALFile.parse(absPath, this._cache, this.getMarkdownFileParser())
+      } else if (hasCodeExt(absPath)) {
+        return await FSALCodeFile.parse(absPath, this._cache)
+      } else {
+        return await FSALAttachment.parse(absPath, this._cache)
+      }
+    }
+
     if (await this.isDir(absPath)) {
       throw new Error(`[FSAL] Cannot load file ${absPath} as it is a directory`)
     }
 
-    if (!await this.isFile(absPath)) {
-      throw new Error(`[FSAL] Cannot load file ${absPath}: Not found`)
-    }
+    throw new Error(`[FSAL] Cannot load file ${absPath}: Not found`)
 
-    if (hasMarkdownExt(absPath)) {
-      const parser = this.getMarkdownFileParser()
-      const descriptor = await FSALFile.parse(absPath, this._cache, parser)
-      return descriptor
-    } else if (hasCodeExt(absPath)) {
-      const descriptor = await FSALCodeFile.parse(absPath, this._cache)
-      return descriptor
-    } else {
-      const descriptor = await FSALAttachment.parse(absPath, this._cache)
-      return descriptor
-    }
   }
 
   /**
@@ -741,7 +737,7 @@ export default class FSAL extends ProviderContract {
    *
    * @throws if the path does not exist
    */
-  public async getDescriptorFor (absPath: string, avoidDiskAccess = true): Promise<AnyDescriptor> {
+  public async getDescriptorFor (absPath: string, avoidDiskAccess: boolean = true): Promise<AnyDescriptor> {
     if (avoidDiskAccess) {
       const cacheHit = this._cache.get(absPath)
       if (cacheHit !== undefined) {
@@ -749,9 +745,9 @@ export default class FSAL extends ProviderContract {
       }
     }
 
-    if (await this.isDir(absPath)) {
+    try {
       return await this.getAnyDirectoryDescriptor(absPath)
-    } else {
+    } catch (err: any) {
       return await this.getDescriptorForAnySupportedFile(absPath)
     }
   }
@@ -815,7 +811,7 @@ export default class FSAL extends ProviderContract {
    * whether a root path is a file or folder. If it is a directory, it will read
    * in the directory and any children recursively to construct a list of every
    * file and folder within `absPath` and return it.
-   * 
+   *
    * NOTE: This function will already exclude dotfiles and ignored directories,
    * so this function is safe to consume in terms of what Zettlr should display.
    *