+++ source/common/modules/markdown-editor/renderers/render-links.ts
@@ -48,12 +48,13 @@ function hideLinkMarkers (view: EditorView): RangeSet<Decoration> {
         } else {
           // It's a regular Markdown Link
           const marks = node.node.getChildren('LinkMark')
+          const label = node.node.getChild('LinkLabel')
 
           // We need at least three LinkMarks for regular links: [, ], and (
           // since the parser will also parse ellipses as Links (a.k.a.
           // reference style links). Alternatively, it needs to have a LinkLabel
           // child node
-          if (marks.length < 3) {
+          if (marks.length < 3 && !label ) {
             return false
           }
 
@@ -61,18 +62,10 @@ function hideLinkMarkers (view: EditorView): RangeSet<Decoration> {
             return false // Empty link title -> would hide the entire link
           }
 
-          if (marks.length === 2) {
-            ranges.push(
-              hiddenDeco.range(marks[0].from, marks[0].to),
-              hiddenDeco.range(marks[1].from, marks[1].to)
-            )
-          } else {
-            ranges.push(
-              hiddenDeco.range(marks[0].from, marks[0].to),
-              hiddenDeco.range(marks[1].from, marks[marks.length - 1].to)
-            )
-          }
-
+          ranges.push(
+            hiddenDeco.range(marks[0].from, marks[0].to),
+            hiddenDeco.range(marks[1].from, label ? label.to : marks[marks.length - 1].to)
+          )
         }
       }
     })