+++ CHANGELOG.md
@@ -1,5 +1,75 @@
 # Upcoming
 
+## GUI and Functionality
+
+- **Feature**: You can now collapse YAML frontmatters and Pandoc divs (#6115).
+- **Feature**: Zettlr now remembers folded regions as long as the window remains
+  open, even when you close a document (#6115).
+- **Change**: Due to a configuration change, downgrading Zettlr (which we
+  generally discourage) after updating to this version will lead to your
+  workspaces being closed.
+- Fixed XCode Command Line Tools setup dialog on macOS (#5428).
+- Fixed an issue that prevented opening Markdown files with tables that include
+  Pandoc attribute strings when the TableEditor was enabled (#6110).
+- Fixed footnote tooltip rendering (#6107).
+- Fixed an issue preventing Zettlr from starting if the stats file contained
+  errors (#6127).
+- Improved link target extraction logic when following links (#6098).
+- Performance improvements when following links (#6072).
+- Hide reference link labels when previewing Markdown links (#6097).
+- Improved performance on generating the preview-bibliography in the sidebar by
+  collecting citation keys from the document only in the main process (#6068).
+- Improved math parsing to fix certain cases (#6030).
+- Relaxed filename restrictions for loading custom dictionaries (#6126).
+- Improved Pandoc div parsing (#6120).
+- Update `de-DE` translation (#6081).
+- Update `ja-JP` translation (#6086).
+
+## Under the Hood
+
+- Moved the `openPaths` property. The `openPaths` property was a simple array
+  holding all currently loaded root files and workspaces intermingled. This
+  setting stems from the time when Zettlr could not load arbitrary files, and
+  mixing files and folders is generally a bad practice. This property has now
+  been moved into the new namespace `app` in the settings, and divided into
+  `openFiles` and `openWorkspaces`. This will make a few future checks and
+  features easier to implement, as the separation into files and folders is made
+  explicit and not left to the OS. This will allow us to better clean up the
+  property (personal note: I've seen a colleague's config file which was full of
+  non-existing file paths, and I didn't like that). In rare situations, this
+  change may close all your workspaces, in which case you would have to re-open
+  them.
+- New utility function `disambiguateFile` that can be used to retrieve a
+  time-based non-existing path that can be used for backup purposes.
+- Added a new environment variable, `ZETTLR_DISABLE_UPDATE_CHECK`. If that
+  variable is detected during build, this will hard-disable update checks in the
+  application. This can be used by package maintainers to ensure update checks
+  from within the app are disabled, and Zettlr instead must be updated via the
+  package manager. Please **note** that the build script will only check for the
+  *presence* of the variable, not its contents. To ensure you are building with
+  updates enabled, you need to ensure the variable is not set, for example by
+  running `unset ZETTLR_DISABLE_UPDATE_CHECK` before commencing the build. For
+  more information, please see issue #6075. When this environment variable is
+  present during build, the resulting binary will exhibit this behavior:
+  - The "Check for updates" menu item will not be present.
+  - The "Updates" and "Beta releases" preferences will not be present. Instead,
+    the updates-setting will be replaced by an informative message informing
+    that updates have been disabled for the app.
+  - The update check itself in the update provider will never run.
+  - Zettlr will include an environment variable indicating whether updates are
+    disabled (`1`) or enabled (`0`) that is visible within the debug panel.
+- Added a new environment variable, `BUNDLE_PANDOC`. If this variable is set to
+  a value of `0`, this will prevent the build script from downloading and
+  bundling Pandoc during the build stage. This can be used by package
+  maintainers to reduce the package size when there is a better way to provide
+  Pandoc (e.g., by listing it as a dependency for Zettlr). This also means that
+  the build environment does not require the dependencies that are specifically
+  for the download script.
+- fix: The Pandoc Attribute parser does not throw an error on malformed
+  attribute strings anymore, and instead just returns an empty record (#6110).
+
+# 4.0.0
+
 ## Full TableEditor Rewrite
 
 This release contains a full rewrite of the TableEditor. The old implementation